ops/devrequirements.txt
coverage==4.2
django==1.11.5
Fabric3==1.12.post1
fabutils>=0.0.20
factory-boy==2.8.1
mock==2.0.0
pep8==1.7.0
pylint==1.7.4
pytest==3.2.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-sugar==0.9.0
sphinx-refdoc==0.0.15
sphinx_rtd_theme>=0.2.4
tox==2.9.1
wheel==0.29.0
